+++ main.client.js
@@ -1,16 +1,23 @@
+import './lib/extension.js';
 import './lib/links/extension.js';
-import './lib/query/extension.js';
 import './lib/query/reducers/extension.js';
-import './lib/namedQuery/_extension.js'; //deprecated
 
 export {
     default as createQuery
-} from './lib/query/createQuery.js';
-
-export {
-    default as createNamedQuery
-} from './lib/namedQuery/_createNamedQuery.js'; //deprecated
+} from './lib/createQuery.js';
 
 export {
     default as prepareForProcess
 } from './lib/query/lib/prepareForProcess';
+
+export {
+    default as Query
+} from './lib/query/query.client';
+
+export {
+    default as NamedQuery
+} from './lib/namedQuery/namedQuery.client';
+
+export {
+    default as compose
+} from './lib/compose';