+++ main.server.js
@@ -1,22 +1,33 @@
+import './lib/extension.js';
+import './lib/aggregate';
 import './lib/exposure/extension.js';
 import './lib/links/extension.js';
-import './lib/query/extension.js';
 import './lib/query/reducers/extension.js';
 import './lib/namedQuery/expose/extension.js';
-import './lib/namedQuery/_extension.js'; //deprecated
+import NamedQueryStore from './lib/namedQuery/store';
+import LinkConstants from './lib/links/constants';
+
+export {
+    NamedQueryStore,
+    LinkConstants
+}
 
 export {
     default as createQuery
-} from './lib/query/createQuery.js';
-
-export {
-    default as createNamedQuery
-} from './lib/namedQuery/_createNamedQuery.js'; //deprecated
+} from './lib/createQuery.js';
 
 export {
     default as Exposure
 } from './lib/exposure/exposure.js';
 
 export {
-    default as getDocumentationObject
-} from './lib/documentor/index.js';
+    default as MemoryResultCacher
+} from './lib/namedQuery/cache/MemoryResultCacher';
+
+export {
+    default as BaseResultCacher
+} from './lib/namedQuery/cache/BaseResultCacher';
+
+export {
+    default as compose
+} from './lib/compose';